francophonic/sources/handmade-dictionaries/pyproject.toml
[tool.poetry]
name = "processor"
version = "0.1.0"
description = ""
authors = ["Andre Popovitch <andre@popovit.ch>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3"
requests = "^2.24"
beautifulsoup4 = "^4.9"
google-cloud-translate = "^3.0"
google-cloud-texttospeech = "^2.2"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
process = "processor.all_runner:main"
template = "processor.handmade_dictionary_template_generator:main"
conjugate = "processor.get_conjugations:main"
